ai_agent/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
redis>=4.6.0
motor>=3.1.1
PyYAML>=6.0
